+++ Resources/dungeon-32x32.tsx
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tileset name="dungeon-32x32" tilewidth="32" tileheight="32" tilecount="5" columns="5">
+ <image source="dungeon-32x32.png" width="160" height="32"/>
+</tileset>